Fixed directory issues
- Added descriptions to commands in plugin.yml
