+++ src/main/java/de/moritzmcc/listener/MoveListener.java
@@ -1,0 +1,9 @@
+package de.moritzmcc.listener;
+
+import org.bukkit.event.Listener;
+
+public class MoveListener implements Listener {
+
+
+
+    }
