--- src/main/java/de/moritzmcc/training/listener/MoveListener.java
@@ -1,14 +1,0 @@
-package de.moritzmcc.training.listener;
-
-import de.moritzmcc.training.damager.Damager;
-import de.moritzmcc.training.util.Area;
-import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerMoveEvent;
-
-public class MoveListener implements Listener {
-
-
-
-    }
