+++ src/main/resources/plugin.yml
@@ -4,12 +4,17 @@ main: de.moritzmcc.training.Main
 api-version: '1.20'
 commands:
   adddamager:
+    description: Create a new damager.
 
   deletedamager:
+    description: Remove a damager.
 
   editdamager:
+    description: Edit a damager.
 
   redirectdamager:
+    description: Redirect a damager.
 
   reloadholograms:
+    description: Reload the holograms.
